+++ Change_log.md
@@ -1,5 +1,6 @@
 # [main](https://github.com/jzhang-github/AGAT)
-- Import `pymatgen` module when necessary. See [agat/data/AtomicFeatures.py](agat/data/AtomicFeatures.py#L44)
+- Import `pymatgen` module when necessary. See [agat/data/AtomicFeatures.py](agat/data/AtomicFeatures.py#L44). Recovered.
+- Specify device when building graphs. See [agat/app/GatApp.py](agat/app/GatApp.py#L69), [agat/data/data.py](agat/data/data.py#L79-83)
 
 # [v7.12](https://github.com/jzhang-github/AGAT/tree/v7.12)